Add berries, deer spawner, and deer
Removed the distinction between resource type and name, resources are now
a single datapoint for simplicity. Hooray for scope cutting!

There were also a bunch of bugfixes in this commit as part of that transition,
some missing state intialization, some issues with state management for
creatures, and there's better debug drawing now for creatures.